feat: 5장 configmap 작업

+++ 05_Manage_Access_1/02_user_permission_configmap/03_attach_policy.sh
@@ -13,6 +13,14 @@ if [ ! -d "tmp" ]; then
     mkdir -p tmp
 fi
 
+# Check if policy exists
+EXISTING_POLICY=$(aws iam get-policy --policy-arn arn:aws:iam::${AWS_ACCOUNT_ID}:policy/${POLICY_NAME} ${PROFILE_STRING} 2>/dev/null)
+
+if [ ! -z "$EXISTING_POLICY" ]; then
+    echo "Policy ${POLICY_NAME} 가 존재합니다."
+    exit 0
+fi
+
 cat >tmp/eks-edu-user-policy.json <<EOF
 {
     "Version": "2012-10-17",
@@ -26,12 +34,16 @@ cat >tmp/eks-edu-user-policy.json <<EOF
 }
 EOF
 
+echo "${POLICY_NAME} 생성중..."
 # eks user policy 생성
 echo "aws iam create-policy --policy-name ${POLICY_NAME} \\
 	--policy-document file://tmp/eks-edu-user-policy.json ${PROFILE_STRING}"
 
 aws iam create-policy --policy-name ${POLICY_NAME} \
 	--policy-document file://tmp/eks-edu-user-policy.json ${PROFILE_STRING}
+
+aws iam wait policy-exists \
+    --policy-arn arn:aws:iam::${AWS_ACCOUNT_ID}:policy/${POLICY_NAME} ${PROFILE_STRING}
 
 # Attach policy to allow EKS access
 echo "aws iam attach-user-policy \\
@@ -41,3 +53,4 @@ echo "aws iam attach-user-policy \\
 aws iam attach-user-policy \
     --user-name eks-edu-user-${IDE_NAME} \
     --policy-arn arn:aws:iam::${AWS_ACCOUNT_ID}:policy/${POLICY_NAME} ${PROFILE_STRING}
+echo "${POLICY_NAME} 생성완료..."
